Implemnted code review changes. Added togle fields for filters and hide/show toogle for operators. Added Date filter options. Added layout display options.
On-behalf-of: @salsadigitalauorg joshua.fernandes@salsadigital.com.au
