--- modules/tide_content_collection/src/Plugin/Field/FieldWidget/ContentCollectionConfigurationWidgetEnhanced.php
@@ -1,639 +1,0 @@
-<?php
-
-namespace Drupal\tide_content_collection\Plugin\Field\FieldWidget;
-
-use Drupal\Core\Extension\ModuleHandlerInterface;
-use Drupal\Core\Field\FieldDefinitionInterface;
-use Drupal\Core\Field\FieldItemListInterface;
-use Drupal\Core\Field\Plugin\Field\FieldWidget\StringTextareaWidget;
-use Drupal\Core\Form\FormStateInterface;
-use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
-use Drupal\tide_automated_listing\SearchApiIndexHelper;
-use Symfony\Component\DependencyInjection\ContainerInterface;
-
-/**
- * Implementation of the content collection configuration enhanced widget.
- *
- * @FieldWidget(
- *   id = "content_collection_configuration_enhanced",
- *   label = @Translation("Content Collection Configuration (Enhanced)"),
- *   field_types = {
- *     "content_collection_configuration"
- *   }
- * )
- */
-class ContentCollectionConfigurationWidgetEnhanced extends StringTextareaWidget implements ContainerFactoryPluginInterface {
-
-  /**
-   * The Search API Index helper.
-   *
-   * @var \Drupal\tide_automated_listing\SearchApiIndexHelper
-   */
-  protected $indexHelper;
-
-  /**
-   * The module handler service.
-   *
-   * @var \Drupal\Core\Extension\ModuleHandlerInterface
-   */
-  protected $moduleHandler;
-
-  /**
-   * The search API index.
-   *
-   * @var \Drupal\search_api\IndexInterface
-   */
-  protected $index;
-
-  /**
-   * {@inheritdoc}
-   */
-  public function __construct($plugin_id, $plugin_definition, FieldDefinitionInterface $field_definition, array $settings, array $third_party_settings, ModuleHandlerInterface $module_handler, SearchApiIndexHelper $index_helper) {
-    parent::__construct($plugin_id, $plugin_definition, $field_definition, $settings, $third_party_settings);
-    $this->moduleHandler = $module_handler;
-    $this->indexHelper = $index_helper;
-    $this->getIndex();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
-    return new static(
-      $plugin_id,
-      $plugin_definition,
-      $configuration['field_definition'],
-      $configuration['settings'],
-      $configuration['third_party_settings'],
-      $container->get('module_handler'),
-      $container->get('tide_automated_listing.sapi_index_helper')
-    );
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function defaultSettings() {
-    return [
-      'schema_validation' => FALSE,
-      'callToAction' => FALSE,
-      'contentTypes' => TRUE,
-      'field_topic' => TRUE,
-      'field_tags' => TRUE,
-    ] + parent::defaultSettings();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function settingsSummary() : array {
-    $summary = parent::settingsSummary();
-
-    $schema_validation = $this->getSetting('schema_validation');
-    $summary[] = $this->t('Validate against the JSON schema: @validation', [
-      '@validation' => $schema_validation ? $this->t('Yes') : $this->t('No'),
-    ]);
-
-    $callToAction = $this->getSetting('callToAction');
-    $summary[] = $this->t('Call To Action Field: @status', [
-      '@status' => $callToAction ? $this->t('Enabled') : $this->t('Disabled'),
-    ]);
-
-    $contentTypes = $this->getSetting('contentTypes');
-    $summary[] = $this->t('Content Types Field: @status', [
-      '@status' => $contentTypes ? $this->t('Enabled') : $this->t('Disabled'),
-    ]);
-
-    $field_topic = $this->getSetting('field_topic');
-    $summary[] = $this->t('Content Topics Field: @status', [
-      '@status' => $field_topic ? $this->t('Enabled') : $this->t('Disabled'),
-    ]);
-
-    $field_tags = $this->getSetting('field_tags');
-    $summary[] = $this->t('Content Tags Field: @status', [
-      '@status' => $field_tags ? $this->t('Enabled') : $this->t('Disabled'),
-    ]);
-
-    $entity_reference_fields = $this->getEntityReferenceFields();
-
-    if (!empty($entity_reference_fields)) {
-      foreach ($entity_reference_fields as $field_id => $field_label) {
-        $field_advanced = $this->getSetting($field_id);
-        $text = $this->t('Field: @status', [
-          '@status' => $field_advanced ? $this->t('Enabled') : $this->t('Disabled'),
-        ]);
-        $summary[] = $field_label['label'] . ' ' . $text;
-      }
-    }
-
-    return $summary;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function settingsForm(array $form, FormStateInterface $form_state) : array {
-    $element = parent::settingsForm($form, $form_state);
-
-    $element['schema_validation'] = [
-      '#type' => 'checkbox',
-      '#title' => $this->t('Validate against the JSON schema'),
-      '#default_value' => $this->getSetting('schema_validation'),
-    ];
-    $element['callToAction'] = [
-      '#type' => 'checkbox',
-      '#title' => $this->t('Enable/Disable Call To Action'),
-      '#default_value' => $this->getSetting('callToAction'),
-    ];
-    $element['contentTypes'] = [
-      '#type' => 'checkbox',
-      '#title' => $this->t('Enable/Disable Content Types'),
-      '#default_value' => $this->getSetting('contentTypes'),
-    ];
-    $element['contentFields']['field_topic'] = [
-      '#type' => 'checkbox',
-      '#title' => $this->t('Enable/Disable Content Topics Field'),
-      '#default_value' => $this->getSetting('field_topic'),
-    ];
-    $element['contentFields']['field_tags'] = [
-      '#type' => 'checkbox',
-      '#title' => $this->t('Enable/Disable Content Tags Field'),
-      '#default_value' => $this->getSetting('field_tags'),
-    ];
-
-    $entity_reference_fields = $this->getEntityReferenceFields();
-
-    if (!empty($entity_reference_fields)) {
-      foreach ($entity_reference_fields as $field_id => $field_label) {
-        $element['contentFields'][$field_id] = [
-          '#type' => 'checkbox',
-          '#title' => $this->t('Enable/Disable') . ' ' . $field_label['label'],
-          '#default_value' => $this->getSetting($field_id),
-        ];
-      }
-    }
-
-    return $element;
-  }
-
-  /**
-   * Get search API index.
-   *
-   * @return \Drupal\search_api\IndexInterface|null|false
-   *   The index, NULL upon failure, FALSE when no index is selected.
-   */
-  protected function getIndex() {
-    if (!$this->index) {
-      // Load and verify the index.
-      /** @var \Drupal\search_api\IndexInterface $index */
-      $index = NULL;
-      $index_id = $this->fieldDefinition->getFieldStorageDefinition()
-        ->getSetting('index');
-      if ($index_id) {
-        $index = $this->indexHelper->loadSearchApiIndex($index_id);
-        if ($index && $this->indexHelper->isValidNodeIndex($index)) {
-          $this->index = $index;
-        }
-      }
-      else {
-        return FALSE;
-      }
-    }
-
-    return $this->index;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function formElement(FieldItemListInterface $items, $delta, array $element, array &$form, FormStateInterface $form_state) {
-    $element = parent::formElement($items, $delta, $element, $form, $form_state);
-
-    // Hide the YAML configuration field.
-    $element['value']['#access'] = FALSE;
-
-    // Load and verify the index.
-    /** @var \Drupal\search_api\IndexInterface $index */
-    $index = $this->getIndex();
-    $index_error = '';
-    if ($index === NULL) {
-      $index_error = $this->t('Invalid Search API Index.');
-    }
-    elseif ($index === FALSE) {
-      $index_error = $this->t('No Search API Index has been selected for this field.');
-    }
-
-    if (!$index) {
-      $element['error'] = [
-        '#type' => 'markup',
-        '#markup' => $index_error,
-        '#prefix' => '<div class="form-item--error-message">',
-        '#suffix' => '</div>',
-        '#allowed_tags' => ['div'],
-      ];
-      return $element;
-    }
-
-    $json = $element['value']['#default_value'];
-    if (!empty($json)) {
-      $json_object = json_decode($json, TRUE);
-      if ($json_object === NULL) {
-        $json_object = [];
-      }
-    }
-
-    $element['title'] = [
-      '#title' => $this->t('Title'),
-      '#type' => 'textfield',
-      '#description' => 'Title displayed above results.',
-      '#default_value' => $json_object['title'] ?? '',
-      '#weight' => 1,
-    ];
-
-    $element['description'] = [
-      '#title' => $this->t('Description'),
-      '#type' => 'textarea',
-      '#description' => 'Description displayed above the results',
-      '#default_value' => $json_object['description'] ?? '',
-      '#weight' => 2,
-    ];
-
-    if ($this->getSetting('callToAction')) {
-      $element['callToAction'] = [
-        '#type' => 'details',
-        '#title' => $this->t('Call To Action'),
-        '#description' => 'A link to another page.',
-        '#open' => TRUE,
-        '#weight' => 3,
-      ];
-      $element['callToAction']['text'] = [
-        '#type'          => 'textfield',
-        '#title'         => $this->t('Text'),
-        '#description'   => $this->t('Display text of the link.'),
-      ];
-      $element['callToAction']['url'] = [
-        '#type'          => 'url',
-        '#title'         => $this->t('Url'),
-        '#description'   => $this->t('A relative or absolute URL.'),
-      ];
-    }
-
-    $configuration = $items[$delta]->configuration ?? [];
-
-    $element['#attached']['library'][] = 'field_group/formatter.horizontal_tabs';
-
-    $element['tabs'] = [
-      '#type' => 'horizontal_tabs',
-      '#tree' => TRUE,
-      '#weight' => 4,
-      '#group_name' => 'tabs',
-    ];
-
-    $this->buildContentTab($items, $delta, $element, $form, $form_state, $configuration, $json_object);
-    $this->buildLayoutTab($items, $delta, $element, $form, $form_state, $configuration, $json_object);
-
-    return $element;
-  }
-
-  /**
-   * Build Content Tab.
-   *
-   * @param \Drupal\Core\Field\FieldItemListInterface $items
-   *   Field items.
-   * @param int $delta
-   *   The current delta.
-   * @param array $element
-   *   The element.
-   * @param array $form
-   *   The form.
-   * @param \Drupal\Core\Form\FormStateInterface $form_state
-   *   The form state.
-   * @param array $configuration
-   *   The YAML configuration of the listing.
-   * @param array $json_object
-   *   The json_object of the listing.
-   */
-  protected function buildContentTab(FieldItemListInterface $items, $delta, array &$element, array &$form, FormStateInterface $form_state, array $configuration = NULL, array $json_object = NULL) {
-
-    $element['tabs']['content'] = [
-      '#type' => 'details',
-      '#open' => TRUE,
-      '#collapsible' => TRUE,
-      '#title' => $this->t('Content'),
-      '#group_name' => 'tabs_content',
-    ];
-
-    if ($this->indexHelper->isNodeTypeIndexed($this->index) && $this->getSetting('contentTypes')) {
-      $element['tabs']['content']['contentTypes'] = [
-        '#type' => 'checkboxes',
-        '#title' => $this->t('Select Content types'),
-        '#options' => $this->indexHelper->getNodeTypes(),
-        '#default_value' => $json_object['internal']['contentTypes'] ?? [],
-        '#weight' => 1,
-      ];
-    }
-
-    if ($this->indexHelper->isFieldTopicIndexed($this->index) && $this->getSetting('field_topic')) {
-      $default_values = $json_object['internal']['contentFields']['field_topic']['values'] ?? [];
-      $field_filter = $this->indexHelper->buildEntityReferenceFieldFilter($this->index, 'field_topic', $default_values);
-      if ($field_filter) {
-        $element['tabs']['content']['field_topic_wrapper'] = [
-          '#type' => 'details',
-          '#title' => 'Select topics',
-          '#open' => FALSE,
-          '#collapsible' => TRUE,
-          '#group_name' => 'tabs_content_filters_field_topic_wrapper',
-          '#weight' => 2,
-        ];
-        $element['tabs']['content']['field_topic_wrapper']['field_topic'] = $field_filter;
-        $element['tabs']['content']['field_topic_wrapper']['operator'] = $this->buildFilterOperatorSelect($json_object['internal']['contentFields']['field_topic']['operator'] ?? 'OR', $this->t('This filter operator is used to combined all the selected values together.'));
-        if (isset($json_object['internal']['contentFields']['field_topic'])) {
-          $element['tabs']['content']['field_topic_wrapper']['#open'] = TRUE;
-        }
-      }
-    }
-
-    if ($this->indexHelper->isFieldTagsIndexed($this->index) && $this->getSetting('field_tags')) {
-      $default_values = $json_object['internal']['contentFields']['field_tags']['values'] ?? [];
-      $field_filter = $this->indexHelper->buildEntityReferenceFieldFilter($this->index, 'field_tags', $default_values);
-      if ($field_filter) {
-        $element['tabs']['content']['field_tags_wrapper'] = [
-          '#type' => 'details',
-          '#title' => 'Select tags',
-          '#open' => FALSE,
-          '#collapsible' => TRUE,
-          '#group_name' => 'tabs_content_filters_field_tags_wrapper',
-          '#weight' => 3,
-        ];
-        $element['tabs']['content']['field_tags_wrapper']['field_tags'] = $field_filter;
-        $element['tabs']['content']['field_tags_wrapper']['operator'] = $this->buildFilterOperatorSelect($json_object['internal']['contentFields']['field_tags']['operator'] ?? 'OR', $this->t('This filter operator is used to combined all the selected values together.'));
-        if (isset($json_object['internal']['contentFields']['field_tags'])) {
-          $element['tabs']['content']['field_tags_wrapper']['#open'] = TRUE;
-        }
-      }
-    }
-
-    $element['tabs']['content']['show_advanced_filters'] = [
-      '#type' => 'checkbox',
-      '#title' => $this->t('Show advanced filters.'),
-      '#description'   => $this->t('Show detailed filters to further limit the overall results.'),
-      '#default_value' => TRUE,
-      '#weight' => 4,
-    ];
-
-    $element['tabs']['content']['advanced_filters'] = [
-      '#type' => 'details',
-      '#title' => $this->t('Advanced filters'),
-      '#open' => TRUE,
-      '#collapsible' => TRUE,
-      '#group_name' => 'tabs_content_advanced_filters',
-      '#weight' => 5,
-      '#states' => [
-        'visible' => [
-          ':input[name="' . $this->getFormStatesElementName('tabs|content|show_advanced_filters', $items, $delta, $element) . '"]' => ['checked' => TRUE],
-        ],
-      ],
-    ];
-
-    // Generate all entity reference filters.
-    $entity_reference_fields = $this->getEntityReferenceFields($items, $delta);
-
-    if (!empty($entity_reference_fields)) {
-      foreach ($entity_reference_fields as $field_id => $field_label) {
-        $default_values = $json_object['internal']['contentFields'][$field_id]['values'] ?? [];
-        $field_filter = $this->indexHelper->buildEntityReferenceFieldFilter($this->index, $field_id, $default_values);
-        if ($field_filter) {
-          $element['tabs']['content']['advanced_filters'][$field_id . '_wrapper'] = [
-            '#type' => 'details',
-            '#title' => $field_label['label'],
-            '#open' => FALSE,
-            '#collapsible' => TRUE,
-            '#group_name' => 'tabs_content_advanced_filters_' . $field_id . '_wrapper',
-          ];
-          $element['tabs']['content']['advanced_filters'][$field_id . '_wrapper'][$field_id] = $field_filter;
-          $element['tabs']['content']['advanced_filters'][$field_id . '_wrapper']['operator'] = $this->buildFilterOperatorSelect($json_object['internal']['contentFields'][$field_id]['operator'] ?? 'OR', $this->t('This filter operator is used to combined all the selected values together.'));
-          if (isset($json_object['internal']['contentFields'][$field_id]['values'])) {
-            $element['tabs']['content']['advanced_filters'][$field_id . '_wrapper']['#open'] = TRUE;
-            $element['tabs']['content']['show_advanced_filters']['#default_value'] = TRUE;
-          }
-        }
-      }
-    }
-
-    // Build extra filters.
-    $extra_filters = $this->moduleHandler->invokeAll('tide_content_collection_extra_filters_build', [
-      $this->index,
-      clone $items,
-      $delta,
-      $json_object['internal']['contentFields'] ?? [],
-    ]);
-    $context = [
-      'index' => clone $items,
-      'delta' => $delta,
-      'filters' => $json_object['internal']['contentFields'] ?? [],
-    ];
-    $this->moduleHandler->alter('tide_content_collection_extra_filters_build', $extra_filters, $this->index, $context);
-    if (!empty($extra_filters) && is_array($extra_filters)) {
-      foreach ($extra_filters as $field_id => $field_filter) {
-        $field_id_index = explode('.', $field_id);
-        $field_id_index = is_array($field_id_index) ? reset($field_id_index) : $field_id;
-        // Skip entity reference fields in extra filters.
-        if (isset($entity_reference_fields[$field_id_index])) {
-          continue;
-        }
-        $index_field = $this->index->getField($field_id_index);
-        if ($index_field) {
-          $element['tabs']['content']['advanced_filters'][$field_id . '_wrapper'] = [
-            '#type' => 'details',
-            '#title' => $index_field->getLabel(),
-            '#open' => FALSE,
-            '#collapsible' => TRUE,
-            '#group_name' => 'filters' . $field_id . '_wrapper',
-          ];
-          $element['tabs']['content']['advanced_filters'][$field_id . '_wrapper'][$field_id] = $field_filter;
-          if (empty($field_filter['#disable_filter_operator'])) {
-            $element['tabs']['content']['advanced_filters'][$field_id . '_wrapper']['operator'] = $this->buildFilterOperatorSelect($json_object['internal']['contentFields'][$field_id]['operator'] ?? 'OR', $this->t('This filter operator is used to combined all the selected values together.'));
-          }
-          unset($field_filter['#disable_filter_operator']);
-          if (isset($json_object['internal']['contentFields'][$field_id]['values'])) {
-            $element['tabs']['content']['advanced_filters'][$field_id . '_wrapper']['#open'] = TRUE;
-            $element['tabs']['content']['show_advanced_filters']['#default_value'] = TRUE;
-          }
-        }
-      }
-    }
-
-  }
-
-  /**
-   * Get all entity reference fields.
-   *
-   * @param \Drupal\Core\Field\FieldItemListInterface $items
-   *   Field items.
-   * @param int $delta
-   *   The current delta.
-   *
-   * @return array
-   *   The reference fields.
-   */
-  protected function getEntityReferenceFields(FieldItemListInterface $items = NULL, $delta = NULL) {
-    $entity_reference_fields = $this->indexHelper->getIndexEntityReferenceFields($this->index, ['nid']);
-    // Allow other modules to remove entity reference filters.
-    $excludes = $this->moduleHandler->invokeAll('tide_content_collection_entity_reference_fields_exclude', [
-      $this->index,
-      $entity_reference_fields,
-      !empty($items) ? clone $items : NULL,
-      $delta,
-    ]);
-    // Exclude the below fields as they are loaded manually.
-    $excludes[] = 'field_topic';
-    $excludes[] = 'field_tags';
-    if (!empty($excludes) && is_array($excludes)) {
-      $entity_reference_fields = $this->indexHelper::excludeArrayKey($entity_reference_fields, $excludes);
-    }
-    return $entity_reference_fields;
-  }
-
-  /**
-   * Build Layout Tab.
-   *
-   * @param \Drupal\Core\Field\FieldItemListInterface $items
-   *   Field items.
-   * @param int $delta
-   *   The current delta.
-   * @param array $element
-   *   The element.
-   * @param array $form
-   *   The form.
-   * @param \Drupal\Core\Form\FormStateInterface $form_state
-   *   The form state.
-   * @param array $configuration
-   *   The YAML configuration of the listing.
-   * @param array $json_object
-   *   The json_object of the listing.
-   */
-  protected function buildLayoutTab(FieldItemListInterface $items, $delta, array &$element, array &$form, FormStateInterface $form_state, array $configuration = NULL, array $json_object = NULL) {
-    $element['tabs']['layout'] = [
-      '#type' => 'details',
-      '#open' => TRUE,
-      '#collapsible' => TRUE,
-      '#title' => $this->t('Layout'),
-      '#group_name' => 'layout',
-    ];
-  }
-
-  /**
-   * Build a filter operator select element.
-   *
-   * @param string $default_value
-   *   The default operator.
-   * @param string $description
-   *   The description of the operator.
-   *
-   * @return string[]
-   *   The form element.
-   */
-  protected function buildFilterOperatorSelect($default_value = 'AND', $description = NULL) {
-    return [
-      '#type' => 'select',
-      '#title' => $this->t('Filter operator'),
-      '#description' => $description,
-      '#default_value' => $default_value ?? 'AND',
-      '#options' => [
-        'AND' => $this->t('AND'),
-        'OR' => $this->t('OR'),
-      ],
-    ];
-  }
-
-  /**
-   * Get the element name for Form States API.
-   *
-   * @param string $element_name
-   *   The name of the element.
-   * @param \Drupal\Core\Field\FieldItemListInterface $items
-   *   Field items.
-   * @param int $delta
-   *   Delta.
-   * @param array $element
-   *   The element.
-   *
-   * @return string
-   *   The final element name.
-   */
-  protected function getFormStatesElementName($element_name, FieldItemListInterface $items, $delta, array $element) {
-    $name = '';
-    foreach ($element['#field_parents'] as $index => $parent) {
-      $name .= $index ? ('[' . $parent . ']') : $parent;
-    }
-    $name .= '[' . $items->getName() . ']';
-    $name .= '[' . $delta . ']';
-    foreach (explode('|', $element_name) as $path) {
-      $name .= '[' . $path . ']';
-    }
-    return $name;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function massageFormValues(array $values, array $form, FormStateInterface $form_state) {
-    $values = parent::massageFormValues($values, $form, $form_state);
-    foreach ($values as $delta => &$value) {
-      $config = [];
-      $config['title'] = $value['title'] ?? '';
-      $config['description'] = $value['description'] ?? '';
-      $config['callToAction']['text'] = ['callToAction']['text'] ?? '';
-      $config['callToAction']['url'] = ['callToAction']['url'] ?? '';
-      $config['internal']['contentTypes'] = $value['tabs']['content']['contentTypes'] ? array_values(array_filter($value['tabs']['content']['contentTypes'])) : [];
-      if (isset($value['tabs']['content']['field_topic_wrapper']['field_topic'])) {
-        foreach ($value['tabs']['content']['field_topic_wrapper']['field_topic'] as $index => $reference) {
-          if (!empty($reference['target_id'])) {
-            $config['internal']['contentFields']['field_topic']['values'][] = (int) $reference['target_id'];
-          }
-        }
-        $config['internal']['contentFields']['field_topic']['operator'] = $value['tabs']['content']['field_topic_wrapper']['operator'] ?? NULL;
-      }
-      if (isset($value['tabs']['content']['field_tags_wrapper']['field_tags'])) {
-        foreach ($value['tabs']['content']['field_tags_wrapper']['field_tags'] as $index => $reference) {
-          if (!empty($reference['target_id'])) {
-            $config['internal']['contentFields']['field_tags']['values'][] = (int) $reference['target_id'];
-          }
-        }
-        $config['internal']['contentFields']['field_tags']['operator'] = $value['tabs']['content']['field_tags_wrapper']['operator'] ?? NULL;
-      }
-
-      $entity_reference_fields = $this->getEntityReferenceFields();
-      foreach ($value['tabs']['content']['advanced_filters'] as $wrapper_id => $wrapper) {
-        $field_id = str_replace('_wrapper', '', $wrapper_id);
-        if (isset($wrapper[$field_id])) {
-          // Entity reference fields.
-          if (isset($entity_reference_fields[$field_id])) {
-            foreach ($wrapper[$field_id] as $index => $reference) {
-              if (!empty($reference['target_id'])) {
-                $config['internal']['contentFields'][$field_id]['values'][] = (int) $reference['target_id'];
-              }
-            }
-          }
-          // Extra fields.
-          else {
-            $config['internal']['contentFields'][$field_id]['values'] = is_array($wrapper[$field_id]) ? array_values(array_filter($wrapper[$field_id])) : [$wrapper[$field_id]];
-            $config['internal']['contentFields'][$field_id]['values'] = array_filter($config['internal']['contentFields'][$field_id]['values']);
-          }
-
-          if (!empty($wrapper['operator'])) {
-            $config['internal']['contentFields'][$field_id]['operator'] = $wrapper['operator'];
-          }
-
-          if (empty($config['internal']['contentFields'][$field_id]['values'])) {
-            unset($config['internal']['contentFields'][$field_id]);
-          }
-        }
-      }
-      $value['value'] = json_encode($config);
-    }
-
-    return $values;
-  }
-
-}
